manager: supervisord backend: completing implementation with first working test

+++ manager/knot_resolver_manager/kresd_controller/supervisord/config.py
@@ -1,6 +1,8 @@
 import os.path
 import signal
 from os import kill
+from pathlib import Path
+from time import sleep
 from typing import List
 
 from jinja2 import Template
@@ -15,12 +17,13 @@ SERVER_SOCK = "/tmp/knot-resolver-manager-supervisord.sock"  # created pseudoran
 
 @dataclass
 class SupervisordConfig:
+    instances: List[str]
     unix_http_server: str = SERVER_SOCK
-    instances: List[str] = []
+    pid_file: str = PID_FILE
 
 
 async def _create_config_file(config: SupervisordConfig):
-    path = os.path.realpath(__file__)
+    path = Path(os.path.realpath(__file__)).parent / "supervisord.conf.j2"
     template = await readfile(path)
     config_string = Template(template).render(config=config)
     await writefile(CONFIG_FILE, config_string)
@@ -28,7 +31,8 @@ async def _create_config_file(config: SupervisordConfig):
 
 async def start_supervisord(config: SupervisordConfig):
     await _create_config_file(config)
-    await call(f'supervisord --configuration="{CONFIG_FILE}" --pidfile="{PID_FILE}"', shell=True)
+    await call(f'supervisord --configuration="{CONFIG_FILE}"', shell=True)
+    sleep(1)
 
 
 async def stop_supervisord():
@@ -39,20 +43,23 @@ async def stop_supervisord():
 
 async def update_config(config: SupervisordConfig):
     await _create_config_file(config)
-    await call(f'supervisorctl -s "{SERVER_SOCK}" update')
+    await call(f'supervisorctl -c "{CONFIG_FILE}" update', shell=True)
 
 
 async def restart(id_: str):
-    await call(f'supervisorctl -s "{SERVER_SOCK}" restart {id_}')
+    await call(f'supervisorctl -c "{CONFIG_FILE}" restart {id_}', shell=True)
 
 
 async def is_supervisord_available() -> bool:
-    i = await call("supervisorctl -h", shell=True)
-    i += await call("supervisord -h", shell=True)
+    i = await call("supervisorctl -h > /dev/null", shell=True, discard_output=True)
+    i += await call("supervisord -h > /dev/null", shell=True, discard_output=True)
     return i == 0
 
 
 async def is_supervisord_running() -> bool:
+    if not Path(PID_FILE).exists():
+        return False
+
     pid = int(await readfile(PID_FILE))
     try:
         kill(pid, 0)
